Improve Message to be more correct.
* Allow responses to HEAD requests with body size.
* Enforce informational messages without body.

Improved documentation, restructured the Message state machine.
Now if a method of the machine is called it will move the FSM
to another state and always the same state except if an error
occurs. For this reason the new FSM is more understandable.
Also it occupies 8 bytes less of storage.